+++ server/src/main/java/org/opensearch/cluster/metadata/MappingMetadata.java
@@ -32,29 +32,19 @@
 
 package org.opensearch.cluster.metadata;
 
-import org.opensearch.OpenSearchParseException;
-import org.opensearch.cluster.AbstractDiffable;
 import org.opensearch.cluster.Diff;
+import org.opensearch.cluster.metadata.core.AbstractMappingMetadata;
 import org.opensearch.common.annotation.PublicApi;
 import org.opensearch.common.compress.CompressedXContent;
-import org.opensearch.common.xcontent.XContentFactory;
-import org.opensearch.common.xcontent.XContentHelper;
-import org.opensearch.core.common.bytes.BytesReference;
-import org.opensearch.core.common.io.stream.BufferedChecksumStreamOutput;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
-import org.opensearch.core.common.io.stream.VerifiableWriteable;
-import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.index.mapper.DocumentMapper;
-import org.opensearch.index.mapper.MapperService;
 
 import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.util.Collections;
 import java.util.Map;
-import java.util.Objects;
 
-import static org.opensearch.common.xcontent.support.XContentMapValues.nodeBooleanValue;
+import static org.opensearch.cluster.metadata.MetadataUtils.SINGLE_MAPPING_NAME;
 
 /**
  * Mapping configuration for a type.
@@ -62,144 +52,64 @@ import static org.opensearch.common.xcontent.support.XContentMapValues.nodeBoole
  * @opensearch.api
  */
 @PublicApi(since = "1.0.0")
-public class MappingMetadata extends AbstractDiffable<MappingMetadata> implements VerifiableWriteable {
-    public static final MappingMetadata EMPTY_MAPPINGS = new MappingMetadata(MapperService.SINGLE_MAPPING_NAME, Collections.emptyMap());
+public class MappingMetadata extends AbstractMappingMetadata {
+    public static final MappingMetadata EMPTY_MAPPINGS = new MappingMetadata(SINGLE_MAPPING_NAME, Collections.emptyMap());
 
-    private final String type;
-
-    private final CompressedXContent source;
-
-    private final boolean routingRequired;
-
+    /**
+     * @deprecated use {@link MappingMetadata#MappingMetadata(String, CompressedXContent, Boolean)} instead
+     * @param docMapper Instance of document mapper
+     */
+    @Deprecated
     public MappingMetadata(DocumentMapper docMapper) {
-        this.type = docMapper.type();
-        this.source = docMapper.mappingSource();
-        this.routingRequired = docMapper.routingFieldMapper().required();
+        super(docMapper.type(), docMapper.mappingSource(), docMapper.routingFieldMapper().required());
     }
 
-    @SuppressWarnings("unchecked")
+    public MappingMetadata(String type, CompressedXContent mappingSource, Boolean routingRequired) {
+        super(type, mappingSource, routingRequired);
+    }
+
     public MappingMetadata(CompressedXContent mapping) {
-        this.source = mapping;
-        Map<String, Object> mappingMap = XContentHelper.convertToMap(mapping.compressedReference(), true).v2();
-        if (mappingMap.size() != 1) {
-            throw new IllegalStateException("Can't derive type from mapping, no root type: " + mapping.string());
-        }
-        this.type = mappingMap.keySet().iterator().next();
-        this.routingRequired = isRoutingRequired((Map<String, Object>) mappingMap.get(this.type));
+        super(mapping);
     }
 
-    @SuppressWarnings("unchecked")
     public MappingMetadata(String type, Map<String, Object> mapping) {
-        this.type = type;
-        try {
-            XContentBuilder mappingBuilder = XContentFactory.jsonBuilder().map(mapping);
-            this.source = new CompressedXContent(BytesReference.bytes(mappingBuilder));
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);  // XContent exception, should never happen
-        }
-        Map<String, Object> withoutType = mapping;
-        if (mapping.size() == 1 && mapping.containsKey(type)) {
-            withoutType = (Map<String, Object>) mapping.get(type);
-        }
-        this.routingRequired = isRoutingRequired(withoutType);
-    }
-
-    @SuppressWarnings("unchecked")
-    private boolean isRoutingRequired(Map<String, Object> withoutType) {
-        boolean required = false;
-        if (withoutType.containsKey("_routing")) {
-            Map<String, Object> routingNode = (Map<String, Object>) withoutType.get("_routing");
-            for (Map.Entry<String, Object> entry : routingNode.entrySet()) {
-                String fieldName = entry.getKey();
-                Object fieldNode = entry.getValue();
-                if (fieldName.equals("required")) {
-                    try {
-                        required = nodeBooleanValue(fieldNode);
-                    } catch (IllegalArgumentException ex) {
-                        throw new IllegalArgumentException(
-                            "Failed to create mapping for type [" + this.type() + "]. " + "Illegal value in field [_routing.required].",
-                            ex
-                        );
-                    }
-                }
-            }
-        }
-        return required;
-    }
-
-    public String type() {
-        return this.type;
-    }
-
-    public CompressedXContent source() {
-        return this.source;
-    }
-
-    /**
-     * Converts the serialized compressed form of the mappings into a parsed map.
-     */
-    public Map<String, Object> sourceAsMap() throws OpenSearchParseException {
-        Map<String, Object> mapping = XContentHelper.convertToMap(source.compressedReference(), true).v2();
-        if (mapping.size() == 1 && mapping.containsKey(type())) {
-            // the type name is the root value, reduce it
-            mapping = (Map<String, Object>) mapping.get(type());
-        }
-        return mapping;
-    }
-
-    /**
-     * Converts the serialized compressed form of the mappings into a parsed map.
-     */
-    public Map<String, Object> getSourceAsMap() throws OpenSearchParseException {
-        return sourceAsMap();
-    }
-
-    public boolean routingRequired() {
-        return this.routingRequired;
-    }
-
-    @Override
-    public void writeTo(StreamOutput out) throws IOException {
-        out.writeString(type());
-        source().writeTo(out);
-        // routing
-        out.writeBoolean(routingRequired);
-    }
-
-    @Override
-    public void writeVerifiableTo(BufferedChecksumStreamOutput out) throws IOException {
-        out.writeString(type());
-        source().writeVerifiableTo(out);
-        out.writeBoolean(routingRequired);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        MappingMetadata that = (MappingMetadata) o;
-
-        if (!Objects.equals(this.routingRequired, that.routingRequired)) return false;
-        if (!source.equals(that.source)) return false;
-        if (!type.equals(that.type)) return false;
-
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(type, source, routingRequired);
+        super(type, mapping);
     }
 
     public MappingMetadata(StreamInput in) throws IOException {
-        type = in.readString();
-        source = CompressedXContent.readCompressedString(in);
-        // routing
-        routingRequired = in.readBoolean();
+        super(in);
     }
 
     public static Diff<MappingMetadata> readDiffFrom(StreamInput in) throws IOException {
-        return readDiffFrom(MappingMetadata::new, in);
+        return new MappingMetadataDiff(in);
+    }
+
+    /**
+     * A diff of mapping metadata.
+     *
+     * @opensearch.internal
+     */
+    static class MappingMetadataDiff implements Diff<MappingMetadata> {
+
+        private final Diff<AbstractMappingMetadata> delegate;
+
+        MappingMetadataDiff(StreamInput in) throws IOException {
+            this.delegate = AbstractMappingMetadata.readDiff(in);
+        }
+
+        @Override
+        public MappingMetadata apply(MappingMetadata part) {
+            AbstractMappingMetadata abstractMappingMetadata = delegate.apply(part);
+            return new MappingMetadata(
+                abstractMappingMetadata.type(),
+                abstractMappingMetadata.source(),
+                abstractMappingMetadata.routingRequired()
+            );
+        }
+
+        @Override
+        public void writeTo(StreamOutput out) throws IOException {
+            delegate.writeTo(out);
+        }
     }
 }