+++ server/src/main/java/org/opensearch/common/util/concurrent/OpenSearchExecutors.java
@@ -34,9 +34,7 @@ package org.opensearch.common.util.concurrent;
 
 import org.opensearch.ExceptionsHelper;
 import org.opensearch.common.SuppressForbidden;
-import org.opensearch.common.logging.DeprecationLogger;
 import org.opensearch.common.settings.Setting;
-import org.opensearch.common.settings.Setting.Property;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.node.Node;
@@ -67,47 +65,21 @@ import java.util.function.Function;
  */
 public class OpenSearchExecutors {
 
-    private static final DeprecationLogger deprecationLogger = DeprecationLogger.getLogger(OpenSearchExecutors.class);
-
     /**
      * Setting to manually set the number of available processors. This setting is used to adjust thread pool sizes per node.
+     * @deprecated - use {@link OpenSearchExecutorsUtils#PROCESSORS_SETTING}.
      */
-    public static final Setting<Integer> PROCESSORS_SETTING = new Setting<>(
-        "processors",
-        s -> Integer.toString(Runtime.getRuntime().availableProcessors()),
-        processorsParser("processors"),
-        Property.Deprecated,
-        Property.NodeScope
-    );
+    @Deprecated
+    public static final Setting<Integer> PROCESSORS_SETTING = OpenSearchExecutorsUtils.PROCESSORS_SETTING;
 
     /**
      * Setting to manually control the number of allocated processors. This setting is used to adjust thread pool sizes per node. The
      * default value is {@link Runtime#availableProcessors()} but should be manually controlled if not all processors on the machine are
      * available to OpenSearch (e.g., because of CPU limits).
+     * @deprecated - use {@link OpenSearchExecutorsUtils#NODE_PROCESSORS_SETTING}.
      */
-    public static final Setting<Integer> NODE_PROCESSORS_SETTING = new Setting<>(
-        "node.processors",
-        PROCESSORS_SETTING,
-        processorsParser("node.processors"),
-        Property.NodeScope
-    );
-
-    private static Function<String, Integer> processorsParser(final String name) {
-        return s -> {
-            final int value = Setting.parseInt(s, 1, name);
-            final int availableProcessors = Runtime.getRuntime().availableProcessors();
-            if (value > availableProcessors) {
-                deprecationLogger.deprecate(
-                    "processors_" + name,
-                    "setting [{}] to value [{}] which is more than available processors [{}] is deprecated",
-                    name,
-                    value,
-                    availableProcessors
-                );
-            }
-            return value;
-        };
-    }
+    @Deprecated
+    public static final Setting<Integer> NODE_PROCESSORS_SETTING = OpenSearchExecutorsUtils.NODE_PROCESSORS_SETTING;
 
     /**
      * Returns the number of allocated processors. Defaults to {@link Runtime#availableProcessors()} but can be overridden by passing a
@@ -115,9 +87,11 @@ public class OpenSearchExecutors {
      *
      * @param settings a {@link Settings} instance from which to derive the allocated processors
      * @return the number of allocated processors
+     * @deprecated - use {@link OpenSearchExecutorsUtils#allocatedProcessors(Settings)}.
      */
+    @Deprecated
     public static int allocatedProcessors(final Settings settings) {
-        return NODE_PROCESSORS_SETTING.get(settings);
+        return OpenSearchExecutorsUtils.allocatedProcessors(settings);
     }
 
     public static PrioritizedOpenSearchThreadPoolExecutor newSinglePrioritizing(