+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased 3.x]
 ### Added
+- Created a new lib metadata-common and separated out IndexMetadata with required dependencies from the server ([#19010](https://github.com/opensearch-project/OpenSearch/pull/19010))
 
 ### Changed